Consolidate benchmark tests in one module
